Simplify how Travis calls Gradle
There is no need to call the wrapper task with a version here as we
already use the wrapper script, and that one will bootstrap the Gradle
version as specified in gradle/wrapper/gradle-wrapper.properties.

Also, there is no need to call the clean task as each Travis build
starts from scratch.