Fix error when running single test in kotlintest-tests (#371)
Currently, `AutoCloseListener` is registered project wide, but closables it checks are only closed when `AutoCloseTest` is run. This is fine if you run the entire test suite, but when running an individual test, it causes the listener's assertions to fail.

This PR changes `AutoCloseTest` so that the closables are considered closed unless that test has actually run.