Add soft assertions (#373)
* Add soft assertions

This PR adds the ability to perform multiple assertions in a block without the test stopping on the first failed assertion. The `verifyAll` function takes a lambda, and any assertion failures inside the block will be gathered, then thrown when the block exits.

Many other testing framework have some version of this functionality:

* [AssertJ](https://joel-costigliola.github.io/assertj/assertj-core-features-highlight.html#soft-assertions)
* [TestNG](https://jitpack.io/com/github/cbeust/testng/master/javadoc/org/testng/asserts/SoftAssert.html)
* [Spock](http://spockframework.org/spock/docs/1.1/all_in_one.html#_what_s_new_in_this_release_4)

The implementation works by changing the `should*` dsl functions to send their failures to an `ErrorCollector`, which will gather the errors when inside a `verifyAll` block, or throw the error immediately otherwise. The collected errors are held in thread-local storage to prevent race conditions in the unlikely case that multiple threads invoke `verifyAll` concurrently.

Existing assertion functions that take a lambda often rely on throwing an exception for control flow, and so will continue to throw immediately instead of collecting their errors. This includes `eventually`, `shouldThrow*`, `shouldTimeout`, and property and table testing.

* Rename verifyAll -> assertSoftly