Add documentation to Date matchers (#420)
* Adds documentation for LocalDate.shouldHaveSameYearAs matcher

Work in progress for https://github.com/kotlintest/kotlintest/issues/410, this commit adds the first documentation as an example to be followed for the other methods.

* WIP: Adds documentation to Date's Year comparison

This is a Work in Progress. DO NOT MERGE YET

Some Date Matchers are missing documentation. This commit adds some documentation, for the Year matching of LocalDate, LocalDateTime, OffsetDateTime and ZonedDateTime

Solves https://github.com/kotlintest/kotlintest/issues/410

* WIP: Adds documentation to Date's Month comparison

This is a Work in Progress. DO NOT MERGE YET

Some Date Matchers are missing documentation. This commit adds some documentation, for the Month matching of LocalDate, LocalDateTime, OffsetDateTime and ZonedDateTime

Solves https://github.com/kotlintest/kotlintest/issues/410

* WIP: Adds documentation to Date's Day comparison

This is a Work in Progress. DO NOT MERGE YET

Some Date Matchers are missing documentation. This commit adds some documentation, for the Day matching of LocalDate, LocalDateTime, OffsetDateTime and ZonedDateTime

Solves https://github.com/kotlintest/kotlintest/issues/410

* WIP: Adds documentation to Date's before comparison

This is a Work in Progress. DO NOT MERGE YET

Some Date Matchers are missing documentation. This commit adds some documentation, for the "is Before" matching of LocalDate, LocalDateTime, OffsetDateTime and ZonedDateTime

Solves https://github.com/kotlintest/kotlintest/issues/410

* WIP: Updates documentation to remove samples

As discussed in this pull request, Dokka's samples, although pretty, won't work currently and will create some maintenance overhead.

To fix these potential issue, this commit removes Dokka's sample and adds them directly to the documentation.

Handles #410, but still a work in progress.

* WIP: Adds documentation to Date's after comparison

This is a Work in Progress. DO NOT MERGE YET

Some Date Matchers are missing documentation. This commit adds some documentation, for the "is After" matching of LocalDate, LocalDateTime, OffsetDateTime and ZonedDateTime

Solves https://github.com/kotlintest/kotlintest/issues/410

* WIP: Adds documentation to Date's within comparison

This is a Work in Progress. DO NOT MERGE YET

Some Date Matchers are missing documentation. This commit adds some documentation, for the "is within" matching of LocalDate, LocalDateTime, OffsetDateTime and ZonedDateTime

Solves https://github.com/kotlintest/kotlintest/issues/410

* Formats comments to match Kotlin code style

Jetbrains Kotlin Code Style Guide states that comments should have a space between // and the actual comment. This commit fixes that.

More information can be found in https://kotlinlang.org/docs/reference/coding-conventions.html#formatting
"Put a space after //: // This is a comment"

This is the final commit that handles https://github.com/kotlintest/kotlintest/issues/410

* Fixes wrong LocalDateTime reference instead of ZonedDateTime

Small bugfix in the documentation, in which there was a reference to LocalDateTime where it should be ZonedDateTime.

Handles https://github.com/kotlintest/kotlintest/issues/410

* Fixes wrong samples reference instead of OffsetDateTime

Small bugfix in the documentation, in which there was a reference to an inexistent sample where it should be OffsetDateTime.

Handles https://github.com/kotlintest/kotlintest/issues/410

* Fixes wrong OffsetDateTime reference instead of LocalDate

Small bugfix in the documentation, in which there was a reference to OffsetDateTime where it should be LocalDate.

Handles https://github.com/kotlintest/kotlintest/issues/410

* Fixes wrong LocalDate reference instead of LocalDateTime

Small bugfix in the documentation, in which there was a reference to LocalDate where it should be LocalDateTime.

Handles https://github.com/kotlintest/kotlintest/issues/410

* Fixes type OffsetDateTimes instead of OffsetDateTime

Handles https://github.com/kotlintest/kotlintest/issues/410