Implements tests for Double Matchers (#451)
* Implements tests for Double Matchers

Given the necessity to continually improve the code, having unit tests guarantees that nothing will break when a change is made. This commit focuses on all Double Matchers, refactoring their tests to cover a lot more scenarios and corner cases.

One problem that this commit introduces is that there are way too many lines for a single class. This exposes a code smell on the matchers.kt file, in which ALL the double matchers are placed. This is bad, as each responsibility should be refactored to it's own class. This correction is possible when issue #450 is solved.

Solves https://github.com/kotlintest/kotlintest/issues/430

* Removes usage of onDoubleRange { double, _ } to avoid unnecessary calls

Having a underscored parameter to a lambda call (_) in this case creates a lot of overhead, and it's not necessary to do that, as the second parameter is not being used.

This commit changes all calls to onDoubleRange that have a ignored parameter to the matcher call

Part of #430

* Modifies testing to include a very low double

This commit aims to improve coverage by including a ridiculously small double 1e300 to the list of test values. To do that, that value was included along some other values in the constants of Double Generator, and according tests were fixed to do it.

* Changes greater and less calculations to use ULP

As discussed in the pull request #451, trying to increase or decrease the number by any amount isn't always correct if the number approaches some extreme values.

To fix it, one might want to use ULP instead of actual double values, and this commit fixes it.

* Formatted files for Kotlin Style Guide and decreased amount of ULP used

As it's always a good practice, the file got formatted to fit Kotlin's Default Style Guide.
Besides that, a minor adjustment was made to decrease the amount of ULPs used to increase or decrease a double value, reducing the amount of numbers that might lead to infinity and improving the overall quality of the test

* Fixed tests that depended on Double Generator, as it has more values now

A change to Double Generator modified the behavior of this test, so it was modified to reflect these changes.