Improves Failures implementation (#465)
* Improves Failures implementation

As it was, the Failures class would fail to filter classes that are neither from `sun.reflect.*` nor from `java.*`. This commit aims to create a generic implementation that correctly filters stacktraces.

The implemented behavior observes a pattern to clean the stacktrace:

- There may be elements that are not from Kotlintest and are not from User Classes, such as Reflection elements or Java elements (but may be others). These elements are removed.
- Then the code tries to find elements that are indeed from Kotlintest, such as runners and other internals
- After this, it's expected for user classes to be in the stacktrace, so the algorithm stops and returns the stacktrace

By observation of the FailuresTest class, one may believe that only classes from `io.kotlintest` and `sun.reflect` would be in the way of user classes, but that's not the case, and this commit fixes it.

* Changes `shouldRemoveKotlintestElementsFromStacktrace` to var instead of val

Users may want to change this property dynamically, at runtime, for whichever reason. Changing it to a `var` won't impact anything else, so it's good to give users this freedom.

(cherry picked from commit 7b17e627d494b61cc306921690e860bc2dcebf1c)