Refactors Throwable Handling to it's own package, fixing bugs
This commit moves ThrowableHandling to their own package, as it's a pattern KT is following for a while. While moving the classes and updating the package, this commit also adds documentations to all use cases, and add some extra cases that `shouldThrowUnit`, such as `shouldThrowAnyUnit`.

This commit deprecates usage, that should be removed from the library in a future. For now, code will still compile.

By changing the order in which verifications are made, this commit also fixes https://github.com/kotlintest/kotlintest/issues/479, and adds a test for every possible scenario, stabilizing and reviewing these methods