Fix 479 - wrong assertion on shouldThrow (#484)
* Refactors Throwable Handling to it's own package, fixing bugs

This commit moves ThrowableHandling to their own package, as it's a pattern KT is following for a while. While moving the classes and updating the package, this commit also adds documentations to all use cases, and add some extra cases that `shouldThrowUnit`, such as `shouldThrowAnyUnit`.

This commit deprecates usage, that should be removed from the library in a future. For now, code will still compile.

By changing the order in which verifications are made, this commit also fixes https://github.com/kotlintest/kotlintest/issues/479, and adds a test for every possible scenario, stabilizing and reviewing these methods

* Moves Throwable Handling to io.kotlintest package

As discussed in the pull request, there won't be any problem to keep these functions in the `io.kotlintest` package, as they're not too many. This also will avoid issues with the user, as they won't have to change anything for the bugs to be fixed and the code to still work.

* Changes `shouldFail` to work with AssertionError only and improves it's documentation

* Reworded documentation to say that assignments are statements

* Changes Failures implementation to delegate to shouldThrow<AssertionError>

* Fixes broken tests

Some test scenarios broke. They were depending on a bug to work (reported in #479), and they weren't testing correctly. When the bug was gone, the tests started to fail, but fixing their test forces them to work again.