Removes Kotlin_Version override documentation
Pull request #502 updated the README.md documentation to include a workaround to Kotlintest not using Kotlin's latest version. However, this change won't be necessary when Kotlin is updated in version 3.2.0 of Kotlintest.

This commit aims to put this reversion in mind now, so that we don't forget about it later.