Add custom exception parsing for reflection call
When a test fails from an AnnotationSpec, it was wrapped in an InvocationTargetException, due to it being called via reflection. Because of it, the test failure stacktrace was confusing, as it wouldn't display user's exception directly.

As explained in https://stackoverflow.com/questions/6020719/what-could-cause-java-lang-reflect-invocationtargetexception , this commit attempts to unwrap an exception if it's thrown by a reflection call context, by returning it's cause instead.

Fixes #539