Enables test bang on all specs (#606)
Before this, some specs couldn't be marked as banged with a `!`. This happened due to the fact that we check for a exclamation mark at the beginning of a test name, such as "Foo" -> "!Foo" would be banged in FreeSpec.

However, some specs such as `BehaviorSpec` wouldn't allow this to work, as `Given("!Foo")` would be turned into `Given: !Foo`. When we checked for the start of the test name, it wouldn't contain `!`, therefore it wouldn't be ignored.

This commit changes all specs that would fall in that scenario, to move an exclamation mark from the beginning of the test chosen name to the beginning of the test final name (Before Given, for example. `Given("!Foo") -> "!Given: Foo"`), fixing this issue.

I chose to do it this way to keep consistent. "Why did this test get ignored?" would be better visualized when `Given` is prefixed with `!`, and not if we edited the bang mechanism to consider only the test name, and not the prefix.

Fixes #605