Provide range-based numeric generators and javax.time generators #530 (#543)
In the effort to add more customization to users, this commit closes #530. For most use cases, users never expect Doubles such as NaN or INFINITY. Due to this, this commit adds some generators for strictly numeric Doubles, which provide ranges of numbers to be chosen from.

This commit also makes some work in `javax.time` generators, as per #526. The classes `LocalDate`, `LocalDateTime` and `LocalTime` now have generators for them, which were also added to the default generators.

Closes #530 
Closes #526