Allows focusing tests with prefix
Before this, and with commit 31d524513970947442385e849e8fca304dea1681, any focused test that contains a prefix wouldn't be correctly focused, because the focus wouldn't be brought behind the prefix.

As of the forementioned commit already delegated test name creation to `createTestName`, this commit just adapt that behavior to include focus.
Fixes #699