Add suspend alternative to Data Driven Testing (#712)
This commit creates a suspended alternative for the data driven testing methods. It also inlines all possible functions in TableTesting. The same approach was not used for the data driven due to the fact that we're using reflect on the functions to obtain the parameter name, and that makes inlining impossible.

To avoid too many maintenance points, I decided to refactor the functions a little bit, so that the execution could be used both in the normal version and the suspended version, shortening the code a little bit.

However, I couldn't find any alternatives for the tests, so I duplicated every test that tested the data driven functions to use the suspended overload. An important note on this is that the function name is not `forallSuspendOverload`, it's just how I imported it.

Fixes #707