Move String matchers to common
This commit moves String and associated code to the common modules. An important point to notice is that this commit will introduce some breaking changes, as some files were moved together to the common modules.

It was assumed that this is OK, as perhaps KotlinTest these changes are for KotlinTest 4.0, in which some breaking changes are expected.