add the ability for eventually to defaultly handle AssertionErrors
A strong use case for using `eventually` is so that something you are
asserting eventually becomes true. Before this change you would have to
explicitly pass `AssertionError` as an exception to handle. Now it will
do this automatically. This enables you to pass a different exception if
you should so choose.

Fixes #753