Fixes tests passing when before test fails
The problem with this issue was that any exception that happened in either `BeforeTest` or `AfterTest` was being silently ignored due to Coroutine Launch instead of `async`.

This commit fixes that error, and also modifies listener execution to notify `JUnit` and our executor to be able to identify these errors and show that in the interface.

Part of #771