Ensure specs that fail in initialisation fail a Maven build
KotlinTest fails to notify the Maven Surefire plugin properly via the
JUnit platform if a spec throws an exception during initialisation.

This is because whilst `TestEngine` does try to register a spec that
fails during initialisation, it accidentally skips the beforeSpecClass
and afterSpecClass methods on the `TestEngineListener` because it is
wrapped in an `IsolationTestEngineListener` which has guards around
those methods, and the guards always return false because the spec was
never created.

See https://github.com/Mahoney/maven-kotlintest for an example of the
problem.