Publish native artifacts for all supported platforms
Currently, only the linux artifact is being published. This PR updates the workflow to use the maxOS runner to publish linux and macOS artifacts, and the windows runner to publish the windows artifact.

The build number is generated once in the build job, then downloaded in the two publish jobs.