Add Arb chunked extensions and Arb element vararg (#1256)
* Add Arb chunked extensions and Arb element vararg

Also, fix the Arb list and set methods so they use a single sequence to generate
lists rather than constructing a new sequence for each list. A new sequence for
each list breaks for arbs that have some sequence initialization, such as distinct
and regex.

* Change constant to call element

This reinforces that the two methods are functionally equivalent.