Remove duplicate classes (#1319)
This commit fixes two issues that were only noticed when Android was
merging jars.

One of them is JUnit's Testable annotation. We keep that annotation as
a way to execute in IntelliJ (green icon). However, when Android is
using something that depends on JUnit itself, our Testable annotation
(which was placed in JUnit's package as it's hardcoded in IntelliJ)
would clash with the official JUnit's Annotation, and android wouldn't
merge the jars.

To fix this issue we needed to include the official annotation in
Kotest. This commit introduces the smallest possible dependency to make
that work.

The other issue was on some aliases which would end on the same JVM
Class, and thus would break the jar merge as well. This was simpler than
 JUnit's.

 Closes #1312