Remove Android from main Kotest project (#1320)
Android code is being maintained in [Kotest-Android](https://github
.com/kotest/kotest-android) project.

This commit removes everything related to Android from the main codebase
 as it's already present in Android's codebase.

Kotest's documentation will still point to Kotest-Android artifacts and
 Android related documentation, but to ease contributions and CI,
 Android should be kept in a separate project.

This commit also removes our integration to JUnit4. We decided to
integrate only with JUnit 5. If JUnit4 becomes necessary again, it will
likely be for Android UI tests only, and thus it could be in
Kotest-Android as well.