Add a test for Koin `checkModules`
At Koin's documentation, one can test that all the modules are setup correctly. However,
the docs only have a guide on doing that with a JUnit extension, and within a JUnit test:
https://start.insert-koin.io/#/getting-started/testing?id=checking-your-modules

This commit adds a test to ensure Kotest fails that assertion correctly. No implementation is necessary
as it already works out of the box.