Adjust Exhaustive default iterations
This commit aims to change how we calculate the default iterations for a given gen. When dealing with Arbs, it's ok to assume a number, as any iterations can be used for an Arb.

However, when talking about Exhaustives, the default iterations might not be enough for them to execute, and thus they fail.

I changed how we compute that default to assume that Exhaustives should have at least their min iterations as default iterations.

Fixes #1456