Add Spec EnabledIf (#1687)
This commit adds the @EnabledIf annotation for specs. The way that Kotlin handles annotations isn't very fexible, and thus
using this annotation is verbose and repetitive. We must reference a class, as we can't reference just a single method as we do for EnabledIf in tests..

This currently only works on JVM, as does @Ignored.

Closes #1683