Performance improvements for Exceptions on the JVM
For our usecase we have lots of exceptions being created by `shouldBe` and similar. We noticed that a lot of time was being spent in class loading. Since it is unlikely the availability of the JUnit 4/5 Assertion classes will change during runtime we can cache the class and constructor lookup. 

Our test suite now runs in 45 seconds instead of 85, which I would consider a worthwhile improvement.