Added inline version for intanceOf alias (#1838)
Added a version of `instanceOf` that is inlined with a reified generic
type. This was done to remove the need for `::class` from the caller.
This type of function is already implemented for `beInstanceOf` and
`beOfType`. This commit just also adds it to `instanceOf`.

Also updated the tests to test every matcher, even aliases. I don't
suspect anyone would accidentally change the implementation of the
aliased functions, but you never know.

Co-authored-by: David Stocking <dstocking@voceracommunications.com>