Fix withEnvironment() to be case-insensitve on Windows (#2099)
On Windows, environment variables names are case-insensitive. The
OverrideMode in withEnvironment() did not take this into account before.
Instead of making OverrideMode itself conditionally ignore case, simply
let it operate on a map with case-insensitive keys on Windows.

Note that the original comment about using "toMap()" was bogus, as all
consumers of "originalEnvironment" anyway already take an immutable map as
their parameter.