Improve the temp{dir,file} functions
Previously, the logic to use javaClass.name as the default prefix could
have been bypassed by setting the prefix to null. Avoid that by making
javaClass.name the fallback if the prefix is null. This is a fixup for
5e78d68.

Also align the temp{dir,file} functions to both use the new API functions
kotlin.io.path.createTemp{Directory,File}, and avoid a duplicate
toFile() conversion.