Fix result matcher nullability (#2762)
* Make the `Throwable` not-null in `shouldBeFailure`

The first assertion verifies that we have a failure, and as such it is
correct and safe to use the not-null assertion operator when invoking the
`block`, which in turn means the block will never be called with a
nullable `Throwable`, so reflect that in the signature.

* ResultMatchersTest: Simplify `runCatching` calls

There is not need to call `runCatching` on the `Result` class, and in
fact that's rather confusing.

* Make the `block` argument in `shouldBeSuccess` match the `Result` type

Use exactly the same (not-null) type when calling the `block` as used in
the `Result`. Fixes #2759.