Adding support for partial permutations
Adding length parameter (defaults to list size) to the permutations function, providing the opportunity of creating exhaustives of partial permutations/k-permutations.

Also, changed the assertion of the full permutations test from shouldBe to shouldContainExactlyInAnyOrder, as the order in which the permutations come are not (should not be) relevant. This also makes the test less dependent on the underlying implementation.

The empty list check was removed, as an empty list will result in a list of an empty list (as with all permutations of length 0).

This commit does break binary compatibility, due to adding the length parameter.